Add About page and apply the styling
New About page and associated css file were created. A "Poppins" font link was added in both the about.html and index.html for styling purposes. CSS styles were added not only for font but also for alignments and color of headers, subheaders, and links. Fixed the size of the HTML Validator in the index.html for better user experience and consistency. All these changes made to enhance the look and navigability of the website for users.