Update relative links to absolute in HTML files
This commit updates all relative links in about.html and index.html files to absolute links. The changes include URLs for CSS, images and internal site navigation. This is done to improve usability across different parts of the website, ensuring that clicking on these links always take users where they are supposed to go, regardless of the current page they are in.